refactor NFT tests
