+++ test/e2e/tests/nft/mocks.js
